xc-bin: move argument "publish" to its own struct
Clap is panicking with "Argument names must be unique" when the
subcommand "run" is called. The reason is that it shares the argument
"publish" with CreateArgs and it's being flattened.

Move it to its own struct and reuse it in "create" and "run".
Note that it seems to be happening in other places as well but
this commit only fixes the "run" command.