xc: handle kevent syscall interruption
kevent will be interrupted and return EINTR on the arrival of a signal.
When calling truss for example against the xcd child process that is
handling the container it will crash because the kevent_ts will return
Err and unwrap() is immediately called on it.

Check if kevent_ts failed and restart the loop if it failed due to a
EINTR, panic! otherwise.